Move the user info data into a new struct
